Fixing FindFromPublicKeyInfo so that it searches the "Public" NSS database if it doesn't find the requested key in the "Private" NSS database.
This fixes the ownership process because the ownership key is created
in the public database because that needs to happen before the TPM is
owned and available (and it's not really all that sensitive to begin
with).

BUG=chromium-os:15645
TEST=Built a new recovery image, wiped a device with it and verified
that I was able to sign in as a new user and add users and forget
networks.  It also showed me as the owner of the device.

Review URL: http://codereview.chromium.org/7066032

git-svn-id: http://src.chromium.org/svn/trunk/src/crypto@86654 4ff67af0-8c30-449e-8e8b-ad334ec8d88c