Properly pass NSS parameters when initializing the PKCS#11 module on CrOS
When initializing the CHAPS PKCS#11 module in CrOS, properly
pass the NSS parameters to SECMOD_LoadUserModule. This
ensures that the default flags for the default slot to mark
the slot as friendly, which means it is not necessary to
call C_Login before calling any read-only operations. Any
actions that fail in read-only mode will still call C_Login.

BUG=118206, chromium-os:28842
TEST=See bug

Review URL: http://codereview.chromium.org/9963127

git-svn-id: http://src.chromium.org/svn/trunk/src/crypto@131075 4ff67af0-8c30-449e-8e8b-ad334ec8d88c